remove the redundant code
