+++ pkg/vectorize/compare/lessThan_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package notequal
+package compare
 
 import (
 	"fmt"
@@ -23,7 +23,7 @@ import (
 	"testing"
 )
 
-func TestI32Eq(t *testing.T) {
+func TestI32Lt(t *testing.T) {
 	as := make([]int32, 10)
 	bs := make([]int32, 10)
 	for i := 0; i < 10; i++ {
@@ -35,7 +35,7 @@ func TestI32Eq(t *testing.T) {
 	bv := testutil.MakeInt32Vector(bs, nil)
 	cv := testutil.MakeBoolVector(cs)
 
-	err := NumericNotEqual[int32](av, bv, cv)
+	err := NumericLessThan[int32](av, bv, cv)
 	if err != nil {
 		t.Fatal(err)
 		t.Fatalf("should not error.")
@@ -43,12 +43,12 @@ func TestI32Eq(t *testing.T) {
 
 	res := vector.MustTCols[bool](cv)
 	for i := 0; i < 10; i++ {
-		fmt.Printf("%+v != %+v : %v \n", as[i], bs[i], res[i])
-		assert.Equal(t, as[i] != bs[i], res[i])
+		fmt.Printf("%+v < %+v \n", as[i], bs[i])
+		assert.Equal(t, as[i] < bs[i], res[i])
 	}
 }
 
-func TestU32Eq(t *testing.T) {
+func TestU32Lt(t *testing.T) {
 	as := make([]uint32, 10)
 	bs := make([]uint32, 10)
 	for i := 0; i < 10; i++ {
@@ -60,7 +60,7 @@ func TestU32Eq(t *testing.T) {
 	bv := testutil.MakeUint32Vector(bs, nil)
 	cv := testutil.MakeBoolVector(cs)
 
-	err := NumericNotEqual[uint32](av, bv, cv)
+	err := NumericLessThan[uint32](av, bv, cv)
 	if err != nil {
 		t.Fatal(err)
 		t.Fatalf("should not error.")
@@ -68,12 +68,12 @@ func TestU32Eq(t *testing.T) {
 
 	res := vector.MustTCols[bool](cv)
 	for i := 0; i < 10; i++ {
-		fmt.Printf("%+v != %+v : %v \n", as[i], bs[i], res[i])
-		assert.Equal(t, as[i] != bs[i], res[i])
+		fmt.Printf("%+v < %+v \n", as[i], bs[i])
+		assert.Equal(t, as[i] < bs[i], res[i])
 	}
 }
 
-func TestF32Eq(t *testing.T) {
+func TestF32Lt(t *testing.T) {
 	as := make([]float32, 2)
 	bs := make([]float32, 2)
 	for i := 0; i < 2; i++ {
@@ -85,19 +85,19 @@ func TestF32Eq(t *testing.T) {
 	bv := testutil.MakeFloat32Vector(bs, nil)
 	cv := testutil.MakeBoolVector(cs)
 
-	err := NumericNotEqual[float32](av, bv, cv)
+	err := NumericLessThan[float32](av, bv, cv)
 	if err != nil {
 		t.Fatalf("should not error.")
 	}
 
 	res := vector.MustTCols[bool](cv)
 	for i := 0; i < 2; i++ {
-		fmt.Printf("%+v != %+v : %v \n", as[i], bs[i], res[i])
-		assert.Equal(t, as[i] != bs[i], res[i])
+		fmt.Printf("%+v < %+v \n", as[i], bs[i])
+		assert.Equal(t, as[i] < bs[i], res[i])
 	}
 }
 
-func TestF64Eq(t *testing.T) {
+func TestF64Lt(t *testing.T) {
 	as := make([]float64, 2)
 	bs := make([]float64, 2)
 	for i := 0; i < 2; i++ {
@@ -109,24 +109,24 @@ func TestF64Eq(t *testing.T) {
 	bv := testutil.MakeFloat64Vector(bs, nil)
 	cv := testutil.MakeBoolVector(cs)
 
-	err := NumericNotEqual[float64](av, bv, cv)
+	err := NumericLessThan[float64](av, bv, cv)
 	if err != nil {
 		t.Fatalf("should not error.")
 	}
 
 	res := vector.MustTCols[bool](cv)
 	for i := 0; i < 2; i++ {
-		fmt.Printf("%+v != %+v : %v \n", as[i], bs[i], res[i])
-		assert.Equal(t, as[i] != bs[i], res[i])
+		fmt.Printf("%+v < %+v \n", as[i], bs[i])
+		assert.Equal(t, as[i] < bs[i], res[i])
 	}
 }
 
-func TestDec64(t *testing.T) {
+func TestDec64Lt(t *testing.T) {
 	as := make([]int64, 10)
 	bs := make([]int64, 10)
 	cs := make([]bool, 10)
 	for i := 0; i < 10; i++ {
-		as[i] = int64(i)
+		as[i] = int64(i + 5)
 		bs[i] = int64(3 * i)
 	}
 
@@ -134,19 +134,19 @@ func TestDec64(t *testing.T) {
 	bv := testutil.MakeDecimal64Vector(bs, nil, types.T_decimal64.ToType())
 	cv := testutil.MakeBoolVector(cs)
 
-	err := Decimal64VecNe(av, bv, cv)
+	err := Decimal64VecLt(av, bv, cv)
 	if err != nil {
-		t.Fatalf("decimal64 not equal failed")
+		t.Fatalf("decimal64 less than failed")
 	}
 
 	res := vector.MustTCols[bool](cv)
 	for i := 0; i < 10; i++ {
-		fmt.Printf("%+v != %+v : %v \n", as[i], bs[i], res[i])
-		assert.Equal(t, as[i] != bs[i], res[i])
+		fmt.Printf("%+v < %+v \n", as[i], bs[i])
+		assert.Equal(t, as[i] < bs[i], res[i])
 	}
 }
 
-func TestDec128(t *testing.T) {
+func TestDec128Lt(t *testing.T) {
 	as := make([]int64, 10)
 	bs := make([]int64, 10)
 	cs := make([]bool, 10)
@@ -159,14 +159,14 @@ func TestDec128(t *testing.T) {
 	bv := testutil.MakeDecimal128Vector(bs, nil, types.T_decimal128.ToType())
 	cv := testutil.MakeBoolVector(cs)
 
-	err := Decimal128VecNe(av, bv, cv)
+	err := Decimal128VecLt(av, bv, cv)
 	if err != nil {
-		t.Fatalf("decimal128 not equal failed")
+		t.Fatalf("decimal128 less than failed")
 	}
 
 	res := vector.MustTCols[bool](cv)
 	for i := 0; i < 10; i++ {
-		fmt.Printf("%+v != %+v : %v \n", as[i], bs[i], res[i])
-		assert.Equal(t, as[i] != bs[i], res[i])
+		fmt.Printf("%+v < %+v \n", as[i], bs[i])
+		assert.Equal(t, as[i] < bs[i], res[i])
 	}
 }
