--- pkg/vectorize/greatequal/greatEqual.go
@@ -1,101 +1,0 @@
-// Copyright 2022 Matrix Origin
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package greatequal
-
-/*
-#include "mo.h"
-
-#cgo CFLAGS: -I../../../cgo
-#cgo LDFLAGS: -L../../../cgo -lmo -lm
-*/
-import "C"
-
-import (
-	"github.com/matrixorigin/matrixone/pkg/common/moerr"
-	"github.com/matrixorigin/matrixone/pkg/container/nulls"
-	"github.com/matrixorigin/matrixone/pkg/container/types"
-	"github.com/matrixorigin/matrixone/pkg/container/vector"
-	"golang.org/x/exp/constraints"
-	"unsafe"
-)
-
-const (
-	LEFT_IS_SCALAR  = 1
-	RIGHT_IS_SCALAR = 2
-)
-
-func NumericGreatEqual[T constraints.Integer | constraints.Float | bool](xs, ys, rs *vector.Vector) error {
-	xt, yt, rt := vector.MustTCols[T](xs), vector.MustTCols[T](ys), vector.MustTCols[bool](rs)
-	flag := 0
-	if xs.IsScalar() {
-		flag |= LEFT_IS_SCALAR
-	}
-	if ys.IsScalar() {
-		flag |= RIGHT_IS_SCALAR
-	}
-
-	rc := C.Numeric_VecGe(unsafe.Pointer(&rt[0]), unsafe.Pointer(&xt[0]), unsafe.Pointer(&yt[0]), C.uint64_t(len(rt)),
-		(*C.uint64_t)(nulls.Ptr(rs.Nsp)), C.int32_t(flag), C.int32_t(int32(xs.Typ.Oid)))
-	if rc != 0 {
-		return moerr.NewError(moerr.INVALID_ARGUMENT, "invalid input type")
-	}
-	return nil
-}
-
-func dec64PtrToC(p *types.Decimal64) *C.int64_t {
-	return (*C.int64_t)(unsafe.Pointer(p))
-}
-func dec128PtrToC(p *types.Decimal128) *C.int64_t {
-	return (*C.int64_t)(unsafe.Pointer(p))
-}
-
-func Decimal64VecGe(xs, ys, rs *vector.Vector) error {
-	xt := vector.MustTCols[types.Decimal64](xs)
-	yt := vector.MustTCols[types.Decimal64](ys)
-	rt := vector.MustTCols[bool](rs)
-	flag := 0
-	if xs.IsScalar() {
-		flag |= LEFT_IS_SCALAR
-	}
-	if ys.IsScalar() {
-		flag |= RIGHT_IS_SCALAR
-	}
-
-	rc := C.Decimal64_VecGE((*C.bool)(&rt[0]), dec64PtrToC(&xt[0]), dec64PtrToC(&yt[0]),
-		C.uint64_t(len(rt)), (*C.uint64_t)(nulls.Ptr(rs.Nsp)), C.int32_t(flag))
-	if rc != 0 {
-		return moerr.NewError(moerr.OUT_OF_RANGE, "Decimal64 euqal error")
-	}
-	return nil
-}
-
-func Decimal128VecGe(xs, ys, rs *vector.Vector) error {
-	xt := vector.MustTCols[types.Decimal128](xs)
-	yt := vector.MustTCols[types.Decimal128](ys)
-	rt := vector.MustTCols[bool](rs)
-	flag := 0
-	if xs.IsScalar() {
-		flag |= LEFT_IS_SCALAR
-	}
-	if ys.IsScalar() {
-		flag |= RIGHT_IS_SCALAR
-	}
-	rc := C.Decimal128_VecGE((*C.bool)(&rt[0]), dec128PtrToC(&xt[0]), dec128PtrToC(&yt[0]),
-		C.uint64_t(len(rt)), (*C.uint64_t)(nulls.Ptr(rs.Nsp)), C.int32_t(flag))
-	if rc != 0 {
-		return moerr.NewError(moerr.OUT_OF_RANGE, "Decimal128 equal error")
-	}
-	return nil
-}
